Update addpost.js to utilise getCookie function
* Validate that a user is signed in by sending the token. Need to
get the token from the cookie.
* Bug with getCookie function missing final character
