+++ src/main/js/cookie/checkCookie.js
@@ -1,17 +1,6 @@
-import history from '../history'
-import { Redirect } from 'react-router'
-
-function getCookie(cookieName){
-    var decodedcookie = decodeURIComponent(document.cookie);
-    var tokenindex = decodedcookie.indexOf(cookieName+"=");
-    var aftertokenindex = decodedcookie.slice(tokenindex);
-    if(aftertokenindex.indexOf(';')){
-        return aftertokenindex.slice(aftertokenindex.indexOf('=') + 1, aftertokenindex.indexOf(';'))
-    }else{
-        return aftertokenindex.slice(aftertokenindex.indexOf('=') + 1);
-    }}
-
-
+import history from '../history';
+import { Redirect } from 'react-router';
+import { getCookie } from './getCookie';
 
 (function checkCookie() {
     var cookie = getCookie("ACKL_token");
