+++ src/main/js/cookie/getCookie.js
@@ -1,0 +1,16 @@
+import history from '../history'
+import { Redirect } from 'react-router'
+
+
+function getCookie(cookieName){
+    var decodedcookie = decodeURIComponent(document.cookie);
+    var tokenindex = decodedcookie.indexOf(cookieName+"=");
+    var aftertokenindex = decodedcookie.slice(tokenindex);
+    if(aftertokenindex.indexOf(';')){
+        return aftertokenindex.slice(aftertokenindex.indexOf('=') + 1, aftertokenindex.indexOf(';'))
+    }else{
+        return aftertokenindex.slice(aftertokenindex.indexOf('=') + 1);
+    }}
+
+export { getCookie }
+
